Move KZG init from SpecLogicEip4844 to MiscHelpersEip4844
