+++ ethereum/spec/src/test/java/tech/pegasys/teku/spec/logic/versions/eip4844/helpers/MiscHelpersEip4844Test.java
@@ -44,7 +44,7 @@ class MiscHelpersEip4844Test {
 
   private final Spec spec = TestSpecFactory.createMinimalEip4844();
   private final MiscHelpersEip4844 miscHelpersEip4844 =
-      new MiscHelpersEip4844(spec.getGenesisSpecConfig(), KZG.NOOP);
+      new MiscHelpersEip4844(spec.getGenesisSpecConfig().toVersionEip4844().orElseThrow());
 
   @BeforeAll
   public static void setup() {}
